feat: 装備選択用のVCを新規作成
