+++ natto/View/Main/EquipmentSelect/ToppingSelect/ToppingSelectViewController.swift
@@ -1,0 +1,15 @@
+//
+//  ToppingSelectViewController.swift
+//  natto
+//
+//  Created by 佐川 晴海 on 2020/10/13.
+//  Copyright © 2020 佐川　晴海. All rights reserved.
+//
+
+import UIKit
+
+class ToppingSelectViewController: UIViewController {
+    override func viewDidLoad() {
+        
+    }
+}
