+++ natto/View/Main/EquipmentSelect/ToppingSelect/ToppingSelectViewController.xib
@@ -18,33 +18,11 @@
         <view clearsContextBeforeDrawing="NO" contentMode="scaleToFill" id="i5M-Pr-FkT">
             <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="Ztd-Rl-v0G" customClass="CloseButton" customModule="natto" customModuleProvider="target">
-                    <rect key="frame" x="362" y="60" width="36" height="36"/>
-                    <constraints>
-                        <constraint firstAttribute="width" constant="36" id="iI3-WI-caS"/>
-                        <constraint firstAttribute="height" constant="36" id="o7R-sF-x61"/>
-                    </constraints>
-                    <state key="normal" title="Button"/>
-                    <connections>
-                        <action selector="dismissAction:" destination="-1" eventType="touchUpInside" id="ccp-6R-YJK"/>
-                    </connections>
-                </button>
-            </subviews>
             <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
             <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-            <constraints>
-                <constraint firstItem="Ztd-Rl-v0G" firstAttribute="top" secondItem="fnl-2z-Ty3" secondAttribute="top" constant="16" id="lKu-hu-ssP"/>
-                <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="Ztd-Rl-v0G" secondAttribute="trailing" constant="16" id="wyM-NU-Zqa"/>
-            </constraints>
             <point key="canvasLocation" x="-145" y="90"/>
         </view>
     </objects>
-    <designables>
-        <designable name="Ztd-Rl-v0G">
-            <size key="intrinsicContentSize" width="46" height="30"/>
-        </designable>
-    </designables>
     <resources>
         <systemColor name="systemBackgroundColor">
             <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
